requirements.txt
aniso8601==2.0.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-language==1.0.0
googleapis-common-protos==1.5.3
grpcio==1.8.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
newsapi-python==0.0.2
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
wincertstore==0.2
